Create folders for unit and integration tests
